requirements.txt
Flask
python-dotenv
openai
pandas
numpy
matplotlib
joblib
requests
scikit-learn
yfinance